Fixes #76 Move screen-specific key handling into each screen module
